requirements.txt
pytest
pytest-html
pytest-xdist
pytest-dotenv
requests
requests-oauthlib
assertpy
pymysql